requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
requests>=2.31.0
pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
